+++ elevenlabs_rs/CHANGELOG.md
@@ -20,6 +20,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `ListKnowledgeBasesDocs` endpoint
 - `ListDependentAgents` endpoint
 - `DeleteKnowledgeBaseDoc` endpoint
+- A `convai::workspace` mod providing the following endpoints:
+  - `GetSettings` to retrieve Convai settings for the workspace
+  - `UpdateSettings` to update workspace settings
+  - `GetSecrets` to list all workspace secrets
+  - `CreateSecret` to create a new workspace secret
 ### Changed
 - **Breaking**: The `ResponseBody` of `DeleteAgent` now returns a `()` instead of a `StatusResponseBody`
 - **Breaking**: The `used_tools` field name to `tool_ids` of the `PromptConfig` and its type `Option<Vec<UsedTool>>`